+++ src/lib/layouts/SplitPane/SplitPane.styles.ts
@@ -1,23 +1,23 @@
 import styled from "styled-components";
 
+interface ResizablePaneProps {
+  minSize?: string;
+  maxSize?: string;
+}
+
 export const SplitPaneContainer = styled.div`
   display: flex;
   flex-direction: row;
   height: 100%;
 `;
 
-interface ResizablePaneProps {
-  minSize?: string;
-  maxSize?: string;
-}
-
 export const ResizablePane = styled.div<ResizablePaneProps>`
   overflow: auto;
-  min-width: ${(props) => props.minSize || '0'};
-  max-width: ${(props) => props.maxSize || '100%'};
+  min-width: ${({minSize}) => minSize || '0'};
+  max-width: ${({maxSize}) => maxSize || '100%'};
 `;
 
 export const Resizer = styled.div`
-  width: 5px;
+  width: 3px;
   cursor: col-resize;
 `;
